fix rewards decimals. refactor API
